refactor(web): inline types from stream-transport-types into stream-transport
The types file existed to break a circular dependency between
stream-transport and stream-watchdog. TypeScript's import type is
erased at compile time and never creates a runtime cycle, making
the separate file unnecessary. The watchdog now uses import type
from stream-transport directly.

Updates CONVENTIONS.md to document the import type exception for
the no-circular-dependencies rule.
